+++ src/Haskellorls/Formatter/Context.hs
@@ -8,6 +8,7 @@ where
 
 import qualified Data.Text as T
 import Haskellorls.Config.Context
+import qualified Haskellorls.Formatter.Attribute as Attr
 import qualified Haskellorls.Formatter.WrappedText as WT
 import qualified Haskellorls.LsColor as Color
 import qualified Haskellorls.NodeInfo as Node
@@ -16,14 +17,14 @@ context :: Node.NodeInfo -> T.Text
 context = Node.getNodeContext
 
 -- | A node SELinux context formatter for the @no@ parameter of the @LS_COLORS@.
-normalColorizedContext :: Color.LsColors -> Node.NodeInfo -> [WT.WrappedText]
-normalColorizedContext lscolors node = [WT.wrap lscolors getter cont]
+normalColorizedContext :: Color.LsColors -> Node.NodeInfo -> [Attr.Attribute WT.WrappedText]
+normalColorizedContext lscolors node = [Attr.Other $ WT.wrap lscolors getter cont]
   where
     cont = context node
     getter = Color.normal
 
-colorizedContext :: Color.LsColors -> Node.NodeInfo -> [WT.WrappedText]
-colorizedContext lscolors node = [WT.wrap lscolors getter cont]
+colorizedContext :: Color.LsColors -> Node.NodeInfo -> [Attr.Attribute WT.WrappedText]
+colorizedContext lscolors node = [Attr.Other $ WT.wrap lscolors getter cont]
   where
     cont = context node
     getter = Color.lookup $ FileContext cont