+++ src/Haskellorls/Formatter/Icon.hs
@@ -1,20 +1,14 @@
-module Haskellorls.Formatter.Icon
-  ( lookupIcon,
-  )
-where
+module Haskellorls.Formatter.Icon (lookupIcon) where
 
-import qualified Data.Text as T
+import qualified Haskellorls.Formatter.Attribute as Attr
 import qualified Haskellorls.Formatter.WrappedText as WT
 import qualified Haskellorls.LsColor as Color
 import qualified Haskellorls.NodeInfo as Node
 
-lookupIcon :: Node.NodeInfo -> Color.LsIcons -> [WT.WrappedText]
+lookupIcon :: Node.NodeInfo -> Color.LsIcons -> [Attr.Attribute WT.WrappedText]
 lookupIcon node lsicons =
-  [ WT.deserialize icon,
-    WT.deserialize margin
+  [ Attr.Other $ WT.deserialize icon,
+    Attr.Other $ WT.deserialize " "
   ]
   where
     icon = maybe "" Color.unIcon $ node `Color.lookup` lsicons
-
-margin :: T.Text
-margin = " "